[peermgr/syncmgr/server]
- Modified blockIndex in BlockInfo to be uint32 and not uint64
- RequestBlocks in Syncmanager takes an index along with the hash
- modified syncmanager code to pass index along with hash in all methods

+++ pkg/peermgr/blockcache.go
@@ -23,7 +23,7 @@ var (
 // to sort and store block requests
 type BlockInfo struct {
 	BlockHash  util.Uint256
-	BlockIndex uint64
+	BlockIndex uint32
 }
 
 // Equals returns true if two blockInfo objects
